added exception for admins for sfw channel greek

+++ guilds/greekgodx.js
@@ -3,6 +3,7 @@ let functions = require('../functions.js');
 
 client.on('message', async message => {
   if (message.channel.id == '746388677978095748') {
+    if (message.member.permissions.has('ADMINISTRATOR')) return;
     if (message.attachments.size == 0) return functions.deleteMessage(message, true);
 
     let attachment = message.attachments.first();
